requirements.txt
imutils==0.5.2
# Use a prebuilt Windows wheel for dlib to avoid local build
dlib-bin==19.24.6
# Use versions with reliable Windows wheels for Python 3.11/3.12
numpy>=1.24,<2
opencv-python>=4.8,<5
playsound==1.2.2
scipy>=1.10,<2
Flask>=2.0
haversine>=0.6
flask-socketio>=5.3.2
eventlet>=0.33.0